Fixes https://github.com/xamarin/xamarin-android/issues/8050

The [`TimeZoneInfo(byte[] data, string id, bool dstDisabled)`](https://github.com/dotnet/runtime/blob/2fbb7eb1ba014de7200e39f048d1ea74a39f0b61/src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.Unix.cs#L35) constructor on Unix sets internal fields used in `DisplayName`, `StandardName`, and `DaylightName` by querying the timezone database via [`TryPopulateTimeZoneDisplayNamesFromGlobalizationData`](https://github.com/dotnet/runtime/blob/5a24757eab72279dcc28f2359ef5771b279efc59/src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.FullGlobalizationData.Unix.cs#L25). For APIs that invoke that constructor for every timezone (>400 timezones on Android) like [`GetSystemTimeZones()`](https://github.com/dotnet/runtime/blob/2fbb7eb1ba014de7200e39f048d1ea74a39f0b61/src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.cs#L856) this operation is costly. 
On average (on my machine, Mac mini M1, 2020 with 16GB memory running on OS 13.4 using Android Studio Emulator), `TryPopulateTimeZoneDisplayNamesFromGlobalizationData` takes ~66% (719.31ms) of the time `GetSystemTimeZones` takes (~1.091s). More specifically, [`GetFullValueForDisplayNameField`](https://github.com/dotnet/runtime/blob/5a24757eab72279dcc28f2359ef5771b279efc59/src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.FullGlobalizationData.Unix.cs#L118) takes ~93% (669ms) of the time of `TryPopulateTimeZoneDisplayNamesFromGlobalizationData`, with the Interop `GetDisplayName`.

Example timing breakdown
```
A breakdown of the `630.37ms` it took to get `GetFullValueForDisplayNameField` is
`336.975ms` `GetDisplayName(timeZoneId, Interop.Globalization.TimeZoneDisplayNameType.Generic, uiCulture.Name, ref genericName);` 
`107.929ms` `GetDisplayName(timeZoneId, Interop.Globalization.TimeZoneDisplayNameType.GenericLocation, uiCulture.Name, ref genericLocationName);`
`1.6074ms`  `GetDisplayName(GmtId, Interop.Globalization.TimeZoneDisplayNameType.GenericLocation, uiCulture.Name, ref gmtLocationName);`
`54.2908ms` `GetDisplayName(GmtId, Interop.Globalization.TimeZoneDisplayNameType.Generic, uiCulture.Name, ref gmtGenericName);`
`1.1826ms` `StringArrayContains(timeZoneId, s_ZonesThatUseLocationName, StringComparison.OrdinalIgnoreCase)`
`17.178ms` `GetExemplarCityName(timeZoneId, uiCulture.Name);`
`75.2069ms` `uiCulture.CompareInfo.IndexOf(genericName, exemplarCityName, CompareOptions.IgnoreCase | CompareOptions.IgnoreNonSpace) >= 0 && genericLocationName != null`
```

This PR looks to improve the performance of APIs invoking a large number of `TimeZoneInfo` constructors by:
1) Lazily loading the values for `DisplayName`, `StandardName`, and `DaylightName`
2) <s>Ordering timezones by Id in `GetSystemTimeZones()` to avoid perf hit from `DisplayName` call.</s>
3) Direct numerical comparison for known UTC alias to reduce number of calls to String.Equals

--------------

On average (on my machine, Mac mini M1, 2020 with 16GB memory running on OS 13.4 using Android Studio Emulator),
Timing `TimeZoneInfo.GetSystemTimeZones()` with the changes in this PR when sorting based on `DisplayName`:
`GetSystemTimeZones`: ~677ms
`TimeZoneInfo(byte[] data, string id, bool dstDisabled)`: ~115.92ms